requirements.txt
Flask>=3.0.0
pymongo[srv]>=4.5.0
dnspython>=2.3.0
Flask-Login>=0.6.3
numpy>=1.26.0
scikit-learn>=1.3.0
Authlib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
